Do not complain about missing column values.
Previously, if an input transformation function didn't receive
any data for a column that was marked as non-optionsal pg_view
complained. Since it relies on external sources like df and /proc,
which may not deliver some results depending on the version of
coreutils and the Linux kernel, we wanted to catch those cases.

With the separation of those data collectors into subprocesses
we sometimes ask them to deliver results while they are not yet
ready, which results in them delivering the empty list. In this
case, we don't want any complaints about individual missing columns,
since the problem is obviously not in external tools.

Per report and diagnostics from @spritchard.