Do not truncate the query containing the '|' char.
pg_view used to serialize both the column name and value into the string,
delimiting them by the '|' char, and deserialize them before the output in
order to calculate the layout and color of the value and, if necessary, the
header. That obvisouly did not work if the string contained '|' characters,
i.e. for a query doing string appends. Since there is no need for
serialization between the collectors and the output code (they are running on
the same system, and even if they were not, there are better ways), the format
was changed to the named tuple, containing the header name, value and the position
of the header relative to the value.

This fixes another problem - for the headers before the values pg_view used to
errorneously highlight the header if the value reached the warning or critical
threshold.